test/benchmark/requirements.txt
scikit-learn
scikit-allel
numpy
pandas
seaborn
